Initial commit
Ghostscript complained about libpng12.0 not being the right version.
It needed version 30.0. and couldn't find it although libpng12.0 in
/usr/local/lib/ is version 30.0.0. Turns out it used the libpng12.0
that comes with shoes wich indeed is version 24.0.0 (via otool).
Solved this issue by symlinking libpng12 from the shoes installation
to the one in /usr/local/lib

Lots of stuff not working yet!!